Insights generation V2.
